+++ CHANGELOG.md
@@ -476,6 +476,7 @@
 * New `paste` command that allows pasting message content that's quoted.
 * New `cartoonize` command and backend (removes parse5).
 * New `hangman` command, adding an option to get a hint, and ability to end a game (type "stop" in the channel).
+* New `ratelimit` command.
 * New array chunking function. Completely remove the old function.
 * New ratelimit system (both global and command), KhafraClient#load now takes a function parameter as a filter.
 * Add `reload` command to reload a command.